Make the word-gap override relative, and sit it beside its siblings
Character Gap is a FRACTION OF LETTER HEIGHT rather than a pixel count.

A pixel threshold silently changes meaning whenever the picture is scaled: the
same title at proxy resolution has gaps half as wide, so a value tuned at full
res merges every word into its line. Timeline resolution and point size do the
same. Letter height moves with all three, so a ratio holds -- and it is the same
yardstick the automatic path already measures in, so the two controls now share a
unit. A word space is about a quarter to a third of letter height in most faces,
so useful values sit near 0.3.

It is stable rather than bit-identical across a scale change: at half resolution
a one-pixel measurement error is twice as large a fraction of letter height, so a
value sitting exactly on the boundary can still flip. Measured full versus half
res, 0.35 and 0.4 agree while 0.3 differs by one group. Pick a value in the
middle of the working range, not at its edge.

The parameter is a new id, deliberately. Reading a stored pixel value as a ratio
would mean 36 times letter height -- a threshold nothing could ever exceed, so
every title would silently collapse to one word per line. Better to lose the
setting and re-dial it once.

Layout: the three grouping controls now read together, Word Gap then Character
Gap then Character Padding, which is what makes their relationship visible. Word
Gap and Character Gap are two ways to make the same decision, the second winning
when it is not zero; Character Padding acts a level below, building the
characters the other two then group. "Character Detection Padding" loses the
middle word now that it sits inside a group where everything is detection.

+++ src/segmentation.h
@@ -22,20 +22,24 @@ struct DetectParams {
   float wordGapSensitivity = 1.0f;
   int bridgeRadius = 0;          // dilate before labeling, to join script fonts
 
-  // The widest gap, in pixels, that still counts as being INSIDE a word.
-  // Anything wider starts a new word.
+  // The widest gap that still counts as being INSIDE a word, as a FRACTION OF
+  // LETTER HEIGHT. Anything wider starts a new word.
   //
-  // Stated as the gap itself rather than as a dilation radius. Bridging achieves
-  // the same grouping by growing every letter by r, which joins gaps up to 2r --
-  // so the number that control takes is half the distance it is really about,
-  // and it describes the mechanism instead of the decision.
+  // Relative, not pixels. A pixel value silently changes meaning the moment the
+  // proxy resolution changes, or the timeline resolution, or the point size --
+  // the same title at half resolution has half the gaps, so a threshold tuned at
+  // full res splits every letter in proxy. Letter height moves with all three,
+  // so a ratio holds. It is also the same yardstick the automatic path uses.
+  //
+  // A word space is roughly a quarter to a third of letter height in most
+  // typefaces, so useful values sit near 0.3.
   //
   // Unlike bridging, the letters are never fused: every character boundary
   // survives, so Character mode still shows letters and a manual split inside a
   // word is still possible.
   //
   // Zero leaves word breaks to the measured gap statistics.
-  int maxLetterGap = 0;
+  float maxLetterGap = 0.0f;
   GroupMode mode = GroupMode::Word;
 
   // Italic compensation. Letters are measured in a sheared coordinate system,
